box-init: Add chained context switches
Each box can specify a handler and a priority for its initialization
procedure.

This commit provides the mechanism to loop over all boxes and call the
uVisor-lib-provided box init function in the context of those boxes. The
box init function comes from flash and it is the same for each box.

The uVisor-lib-provided box init function is passed a per-box
configuration pointer as an argument. That pointer points to
uVisor-lib-specific data in flash that the uVisor does not care about
(opaque type). The uVisor-lib-provided box init function will take care
of interpreting the configuration pointer as something it knows.