core: Add chained context switches for box init
Each box can specify a handler and a priority for its initialization
procedure.

This commit provides the mechanism to loop over all boxes and call a
default handler in the context of those boxes. The default handler comes
from flash, and is not the box-specific initialization handler.

Instead the default handler is passed the following 4 arguments:

    * Box-specific initialization handler.
    * Box-specific initialization handler priority.
    * Initial stack pointer for that secure box.
    * Stack size for that secure box.

Depending on the OS, the default handler can either dispatch the
box-specific handler immediately or post a callback to it later on. On
CMSIS RTOS, a thread is created (belonging to the box active at the time
of creation) which is tied to that box-specific handler.