context: Add function to get the first state of a box
We never store the initial state (stack pointer) of a secure box.
Instead, we initialize the *current* state of all boxes with the
initial stack pointer, which will change at runtime when multiple nested
context switches happen.

This commit introduces a function to find the first known state for a
secure box. The state is looked up in the state stack. If none is
available for the requested box, then the current state is used, which
is equal to the initial state.