Require explicit declaration of page heap sizes
This removes the global `UVISOR_PAGE_SIZE` define and instead introduces
the `uvisor_get_page_size()` getter function to retrieve the page size
for the tier-2 allocator.

The user is now required to declare the minimum page size by explicitly
stating the required page size and the minimum number of pages in the
heap. This allows applications to specify their requirements, which can
then be checked by linker which raises an error if not enough memory
for the page heap is available.